+++ Cargo.toml
@@ -31,7 +31,7 @@ web-sys = {version = "0.3", optional = true, features = ["console", "Performance
 tiny-keccak = {version = "1.5", optional = true}
 blake2-rfc = {version = "0.2.18", optional = true}
 
-blake2s_simd = {version = "0.5", optional = true}
+blake2s_simd = {version = "0.5"}
 lazy_static = {version = "1", optional = true}
 
 blake2s_const = {version = "0.6", optional = true, path = "./src/plonk/blake2_const/blake2s/"}
@@ -43,14 +43,13 @@ blake2s_const = {version = "0.6", optional = true, path = "./src/plonk/blake2_co
 hex = "*"
 
 [features]
-default = ["multicore"]
+default = []
 multicore = ["crossbeam", "futures/thread-pool"]
 sonic = ["tiny-keccak", "blake2-rfc"]
 gm17 = []
 nolog = []
-plonk = ["blake2s_simd", "lazy_static", "tiny-keccak", "blake2s_const"]
+plonk = ["lazy_static", "tiny-keccak", "blake2s_const"]
 # redshift = ["plonk", "rescue_hash", "poseidon_hash"]
-# plonk = ["blake2s_simd", "lazy_static", "tiny-keccak", "blake2s_const", "rescue_hash", "poseidon_hash"]
-marlin = ["blake2s_simd", "tiny-keccak", "blake2s_const"]
+marlin = ["tiny-keccak", "blake2s_const"]
 wasm = ["web-sys"]
 asm = ["pairing/asm"]
